refactor(assignments): extrai LotteryParticipantsSection e LotteryPreviewSection
Zera o no-giant-component do LotteryDialog (16→15 diags no repo; o arquivo
fica com 212 linhas). A PreviewSection agrupa estado/ações do sorteio no
objeto run para não introduzir booleanos soltos; memberName move junto.

Refs #357
